+++ backend/GastronomyApp.Infrastructure/Migrations/20260915222426_InitialCreate.cs
@@ -105,7 +105,7 @@ namespace GastronomyApp.Infrastructure.Migrations
                     CategoryId = table.Column<Guid>(type: "TEXT", nullable: false),
                     SortOrder = table.Column<int>(type: "INTEGER", nullable: false),
                     IsActive = table.Column<bool>(type: "INTEGER", nullable: false),
-                    ProductionMinutes = table.Column<int>(type: "INTEGER", nullable: true),
+                    ProductionMinutes = table.Column<double>(type: "REAL", nullable: true),
                     IsQueueIndependent = table.Column<bool>(type: "INTEGER", nullable: false)
                 },
                 constraints: table =>
